feat: Sale entity için Configuration yapıldı.
Sale tablosu için veri tabanındaki kolonların özelliklerini etkileyecek birtakım ayarlar yapıldı. Ayrıca fiyat alanı decimal olarak değiştirildi.

+++ Onicorn.CRMApp.API/Onicorn.CRMApp.Entities/Sale.cs
@@ -11,7 +11,7 @@ namespace Onicorn.CRMApp.Entities
     {
         [Key]
         public int Id { get; set; }
-        public int SalesAmount { get; set; }
+        public decimal SalesAmount { get; set; }
         public DateTime SalesDate { get; set; }
         public bool Status { get; set; }
